third_party/WebKit/LayoutTests/fast/performance/longtasktiming.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

async_test(function (t) {
    var observer = new PerformanceObserver(
        t.step_func(function (entryList) {
            var entries = entryList.getEntries();
            /* TODO(panicker): Update after fixing http://crbug.com/651461 */
            for (var i = 0; i < entries.length; i++) {
                assert_equals(entries[i].entryType, "longtask",
                    "entryType expected to be: longtask");
                assert_equals(entries[i].name, "same-origin-self",
                    "name expected to be: same-origin-self");
                assert_greater_than(entries[i].duration, 50,
                    "duration expected to be greater than 50ms threshold");
                assert_equals(entries[i].startTime, Math.floor(entries[i].startTime),
                    "startTime expected to have 1 miilisecond granularity");
            }
            observer.disconnect();
            t.done();
        })
    );
    observer.observe({entryTypes: ["longtask"]});

    /* Generate a slow task */
    var begin = window.performance.now();
    while (window.performance.now() < begin + 51);

}, "Performance longtask entries are observable");

</script>
